Create endpoints module and cargo clippy
