+++ src/http/permissions.rs
@@ -5,7 +5,7 @@ pub fn can_create_database(
     desired_entity: &InstantiatedEntity,
 ) -> bool {
     // An entity/user can only create databases on itself (for now)
-    return authenticated_entity.id == desired_entity.id;
+    authenticated_entity.id == desired_entity.id
 }
 
 pub fn can_query(
@@ -13,5 +13,5 @@ pub fn can_query(
     database: &InstantiatedDatabase,
 ) -> bool {
     // An entity/user can only query its own databases (for now)
-    return authenticated_entity.id == database.entity_id;
+    authenticated_entity.id == database.entity_id
 }
